requirements.txt
pinecone-client==3.1.0
PyPDF2==3.0.1
langchain==0.1.12
langchain_pinecone==0.0.3
langchain_google_genai==0.0.11
google-generativeai==0.4.1
python-dotenv==1.0.1
numpy==1.25.2
pypdf==4.1.0
PyPDF2==3.0.1
openai==1.14.0
tiktoken
streamlit==1.32.2

